zephyr: Aligning source directory tree with apache/mynewt-core
Goal of this commit is to make file structure tree, of this fork of
TinyCBOR, to be more aligned to the upstream:
https://github.com/apache/mynewt-core/tree/master/encoding/tinycbor
