+++ include/tinycbor/cbor.h
@@ -34,10 +34,10 @@
 #include <string.h>
 #include <stdio.h>
 
-#include "cbor_buf_writer.h"
-#include "cbor_buf_reader.h"
-#include "tinycbor-version.h"
-#include "config.h"
+#include "tinycbor/cbor_buf_writer.h"
+#include "tinycbor/cbor_buf_reader.h"
+#include "tinycbor/tinycbor-version.h"
+#include "tinycbor/config.h"
 
 #define TINYCBOR_VERSION            ((TINYCBOR_VERSION_MAJOR << 16) | (TINYCBOR_VERSION_MINOR << 8) | TINYCBOR_VERSION_PATCH)
