+++ include/tinycbor/tinycbor-version.h
