+++ src/cbortojson.c
@@ -30,9 +30,9 @@
 #  define __STDC_LIMIT_MACROS 1
 #endif
 
-#include "cbor.h"
-#include "cborjson.h"
-#include "compilersupport_p.h"
+#include "tinycbor/cbor.h"
+#include "tinycbor/cborjson.h"
+#include "tinycbor/compilersupport_p.h"
 
 #include <float.h>
 #include <inttypes.h>
